models

+++ earthstock/models.py
@@ -5,5 +5,5 @@ from django.db import models
 class Stock(models.Model):
     symbol = models.CharField(max_length = 50, null=False)
     company = models.CharField(max_length = 200, null=False)
-    latitude = models.DecimalField()
-    longitude = models.DecimalField()
+    latitude = models.FloatField()
+    longitude = models.FloatField()
